Allow the Linux daemon to be stopped and started via the web-app and for the PIN to be changed.
This is just the bare-bones plumbing needed before the more useful cross-platform work like getting the auth credentials out of the web app can begin.

The UI needs work, but that's a separate CL.

BUG=None
TEST=Make sure that me2me_virtual_host.py is on your PATH and you should be good to go.


Review URL: http://codereview.chromium.org/9653013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@125893 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 3f01c469931174046154274c10838f19492efe9d

+++ remoting/tools/me2me_virtual_host.py
@@ -209,6 +209,9 @@ or above, please set a non-empty PIN. You can change PIN later using
         print "PINs didn't match. Please try again."
         continue
       break
+    self.set_pin(pin)
+
+  def set_pin(self, pin):
     if pin == "":
       self.host_secret_hash = "plain:"
     else:
@@ -534,6 +537,11 @@ def main():
   parser.add_option("-p", "--new-pin", dest="new_pin", default=False,
                     action="store_true",
                     help="set new PIN before starting the host")
+  parser.add_option("", "--check-running", dest="check_running", default=False,
+                    action="store_true",
+                    help="return 0 if the daemon is running, or 1 otherwise")
+  parser.add_option("", "--explicit-pin", dest="explicit_pin", default=None,
+                    help="set or unset the pin on the command line")
   (options, args) = parser.parse_args()
 
   size_components = options.size.split("x")
@@ -542,6 +550,10 @@ def main():
 
   host_hash = hashlib.md5(socket.gethostname()).hexdigest()
   pid_filename = os.path.join(CONFIG_DIR, "host#%s.pid" % host_hash)
+
+  if options.check_running:
+    running, pid = PidFile(pid_filename).check()
+    return 0 if (running and pid != 0) else 1
 
   if options.stop:
     running, pid = PidFile(pid_filename).check()
@@ -578,6 +590,14 @@ def main():
   host = Host(os.path.join(CONFIG_DIR, "host#%s.json" % host_hash))
   register_host = not host.load_config()
 
+  if options.explicit_pin != None:
+    host.set_pin(options.explicit_pin)
+    host.save_config()
+    running, pid = PidFile(pid_filename).check()
+    if running and pid != 0:
+      os.kill(pid, signal.SIGUSR1)
+    return 0
+
   # Outside the loop so user doesn't get asked twice.
   if register_host:
     host.ask_pin()
@@ -585,7 +605,7 @@ def main():
     host.ask_pin()
     host.save_config()
     running, pid = PidFile(pid_filename).check()
-    if running:
+    if running and pid != 0:
       os.kill(pid, signal.SIGUSR1)
       print "The running instance has been updated with the new PIN."
       return 0
